#168: Fix a bug in worker queueing

+++ patient-journey-client/src/data/workers/hooks.ts
@@ -1,50 +1,51 @@
-import { useEffect, useRef, useState } from 'react'
+import { useCallback, useEffect, useRef, useState } from 'react'
 
 export const useWorker = <D, R>(_Worker: new () => Worker, data: D, initialValue: R): R => {
   const [workerInstance, setWorkerInstance] = useState<Worker>()
   const [result, setResult] = useState<R>(initialValue)
-  const [popQueueToggle, setPopQueueToggle] = useState(false)
 
   let isBusy = useRef(false)
   let popData = useRef<any | undefined>(undefined)
 
-  // Initialize worker
-  useEffect(() => {
-    if (!workerInstance) {
-      const worker = new _Worker()
-      worker.addEventListener('message', (e) => {
-        setResult(e.data)
-        isBusy.current = false
-        setPopQueueToggle((toggle) => !toggle)
-      })
-      setWorkerInstance(worker)
-    } else {
-      return () => {
-        workerInstance.terminate()
-        setWorkerInstance(undefined)
-      }
-    }
-  }, [workerInstance, _Worker])
-
-  // Post data to the worker when data changes
-  useEffect(() => {
-    if (workerInstance && isBusy.current === false) {
-      workerInstance.postMessage(data)
-      isBusy.current = true
-    } else if (workerInstance && data) {
-      // Worker is busy, queue data
-      popData.current = data
-    }
-  }, [workerInstance, data])
-
-  useEffect(() => {
-    if (workerInstance && popData.current && isBusy.current === false && popQueueToggle !== undefined) {
-      // Worker is not busy anymore, catch up with latest data
+  const flushPopData = useCallback(() => {
+    if (workerInstance && popData.current && isBusy.current === false) {
       isBusy.current = true
       workerInstance.postMessage(popData.current)
       popData.current = undefined
     }
-  }, [popQueueToggle, workerInstance])
+  }, [workerInstance])
+
+  // Initialize worker
+  useEffect(() => {
+    if (flushPopData) {
+      if (!workerInstance) {
+        const worker = new _Worker()
+        setWorkerInstance(worker)
+      } else {
+        const handleWorkerResponse = (e: MessageEvent<R>) => {
+          setResult(e.data)
+          isBusy.current = false
+          flushPopData()
+        }
+
+        workerInstance.addEventListener('message', handleWorkerResponse)
+
+        return () => {
+          workerInstance.removeEventListener('message', handleWorkerResponse)
+          workerInstance.terminate()
+          setWorkerInstance(undefined)
+        }
+      }
+    }
+  }, [workerInstance, _Worker, flushPopData])
+
+  // Post data to the worker when data changes
+  useEffect(() => {
+    if (workerInstance) {
+      popData.current = data
+      flushPopData()
+    }
+  }, [workerInstance, data, flushPopData])
 
   return result
 }
